Add page title component
